Following #43456 to seperate os related dependence
